+++ Common/AdvancedMultipleMutexRequest.cs
@@ -7,10 +7,14 @@ using StardewValley.Network;
 using StardewModdingAPI;
 using StardewValley;
 using StardewModdingAPI.Events;
+using Leclair.Stardew.Common.Events;
+using Netcode;
 
 namespace Leclair.Stardew.Common;
 
 public class AdvancedMultipleMutexRequest {
+
+	public static ModSubscriber? Mod = null;
 
 	private IModHelper? Helper;
 	private int Timeout;
@@ -30,7 +34,7 @@ public class AdvancedMultipleMutexRequest {
 
 	//private FarmerCollection FC;
 
-	public AdvancedMultipleMutexRequest(IEnumerable<NetMutex> mutexes, Action? onSuccess = null, Action? onFailure = null, IModHelper? helper = null, int timeout = 500) {
+	public AdvancedMultipleMutexRequest(IEnumerable<NetMutex> mutexes, Action? onSuccess = null, Action? onFailure = null, IModHelper? helper = null, int timeout = 1000) {
 		Helper = helper;
 		Timeout = timeout;
 		OnSuccess = onSuccess;
@@ -151,7 +155,54 @@ public class AdvancedMultipleMutexRequest {
 		if (IsLocked()) {
 			OnSuccess?.Invoke();
 
-		} else { 
+		} else {
+			if (Mod != null) {
+				try {
+#if DEBUG
+					LogLevel level = LogLevel.Debug;
+#else
+					LogLevel level = LogLevel.Trace;
+#endif
+
+					Mod.Log($"Unable to acquire all mutexes within {Timeout} ms. IsHost: {Game1.IsMasterGame}; Multiplayer: {Context.IsMultiplayer}; Mutex state:", level);
+
+					List<string[]> states = new();
+
+					foreach (var mutex in Mutexes) {
+						var owner = Helper?.Reflection?.GetField<NetLong>(mutex, "owner", false)?.GetValue();
+						var request = Helper?.Reflection?.GetField<NetEvent1Field<long, NetLong>>(mutex, "lockRequest", false)?.GetValue();
+						object? _inEvents = request is null ? null : Helper!.Reflection!.GetField<object>(request, "incomingEvents", false)?.GetValue();
+						object? _outEvents = request is null ? null : Helper!.Reflection!.GetField<object>(request, "outgoingEvents", false)?.GetValue();
+
+						int inEvents = _inEvents is null ? -1 : Helper!.Reflection!.GetProperty<int>(_inEvents, "Count", false)?.GetValue() ?? -1;
+						int outEvents = _outEvents is null ? -1 : Helper!.Reflection!.GetProperty<int>(_outEvents, "Count", false)?.GetValue() ?? -1;
+
+						states.Add(new string[] {
+							$"{mutex.GetHashCode()}",
+							$"{mutex.IsLocked()}",
+							$"{mutex.IsLockHeld()}",
+							owner is not null ? $"{owner.Value}" : "---",
+							$"{inEvents}",
+							$"{outEvents}"
+						});
+					}
+
+					string[] headers = new string[] {
+						"ID",
+						"Locked",
+						"LockHeld",
+						"Owner",
+						"inEvents",
+						"outEvents"
+					};
+
+					Mod.LogTable(headers, states, level);
+
+				} catch(Exception) {
+					/* do nothing */
+				}
+			}
+
 			ReleaseLock();
 			OnFailure?.Invoke();
 		}